git-svn-id: svn://123.57.173.36/meijia-parent@1574 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-dao/src/main/java/com/simi/po/dao/record/RecordAssetsMapper.java
@@ -10,7 +10,7 @@ public interface RecordAssetsMapper {
 
     int insert(RecordAssets record);
 
-    int insertSelective(RecordAssets record);
+    Long insertSelective(RecordAssets record);
 
     RecordAssets selectByPrimaryKey(Long id);
